Time: 237 ms (29.11%), Space: 23.4 MB (35.63%) - LeetHub

+++ 15-3sum/15-3sum.cpp
@@ -8,7 +8,7 @@ public:
         
         sort(num.begin(),num.end());
         
-        for(int i=0;i<(int)(num.size()-2);i++){
+        for(int i=0;i<(num.size()-2);i++){
             
             if(i==0 || (i>0 && num[i] !=num[i-1])){
